pyproject.toml
[tool.poetry]
name = "nous-bot"
version = "0.1.0"
description = "A highly intelligent AI discord companion."
authors = ["anthoeknee"]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
httpx = "^0.28.0"
colorlog = "^6.9.0"
alembic = "^1.14.0"
aiosqlite = "^0.20.0"
watchdog = "^6.0.0"
redis = "^5.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
